Change from deploy to bamboo user.
Rather than have a separate deploy user and a generic bamboo user
we're going to deploy the application as the bamboo user.

This keeps things simpler - fewer things to manage, fewer keys,
and generally means that anyone who wants to log onto the box
to look at logs or fiddle with the app can do so via the
bastion host.