Fix: Set correct path for vhost
Since deciding to run the application as the `bamboo` user,
changes have been made to the deployable recipe to permit the
app to be deployed as the bamboo user, but the vhost path was
overlooked.

We should probably consider this for an attribute, but I'm ok
with it being hard-coded at this stage.