Add: Tests for base recipe, especially hardening
The EFG environment has very high requirements for security, so
the base recipe implements a range of hardening features.

These tests are at present pretty primitive, in that they do not
actually test the functionality, but they do at least document it
and check that config is in place to deliver it.