Fix: Add node data for chef-zero search
The database and deployable recipes use Chef search, but on a first
run, there will be no node data.  This provides data that will satisfy
the search for efg::training AND efg::database.

We should probably also store the saved state of the vagrant machine
for the machine after testing.