Add: Use backup cookbook and skyscape S3
This uses the Skyscape S3-compatible storage for backups.  Using
the Ruby backup gem, this takes a dump of the database and then
compresses it before sending it to a named bucket.

Currently the secret key is stored in a databag, and is not
encrypted.  For this reason, we do not check the contents of the
secrets databag into git.  This should use encrypted databags or
attributed in the future.